rle: implement intervals
adds an implementation for the `Interval` structure, which is the basis
of a run-length encoding. intervals are defined similarly to CRoaring --
they are a tuple (value, length) which encode the half-open interval [a,
b). the tuple defines `value` to be `a` and `length` to be `b - a`, or
the size of the interval. in addition to the interval definition, we
provide some helpers to interact with the interval in subsequent
operations