rle: implement `interleaved_binary_search`
implements the `interleaved_binary_search` algorithm from croaring. this
binary search returns a result which is Ok(x) if and only if the given
key is a `value` in the internally maintained vector of arrays. in this
case, `x` is the index of the interval that starts with given key. note
that this means that *only the first element of an interval* will return
Ok(x).  every other element will return Err(y), where `y` is the index
of the interval that will potentially contain the element of a given
index. if the `key` is smaller than the value of the first interval,
then `y` is -1