rle: implement xor
implements the exclusive or operation on run stores. the exclusive or
operation returns a new resultant run store from two given run stores.
the elements of this resultant are precisely the elements that are in
one run store, but not the other.

this commit also adds a `BivariateOrderedIterator`. this data type takes
in two iterators and, upon `next`, yields the smaller of the two `next`
elements from both of these iterators.